mottakerinstitusjonvelger henter oppdateringsfunksjoner selv
